Move more classes out of internal
