+++ api/src/org/labkey/api/data/TransactionFilter.java
@@ -31,13 +31,12 @@ public class TransactionFilter implements Filter
     @Override
     public void destroy()
     {
-
     }
 
     @Override
     public void doFilter(ServletRequest req, ServletResponse resp, FilterChain chain) throws IOException, ServletException
     {
-        // Is it worth creating another filter for this?  It is in the spirit of per-request resource tracking.
+        // Is it worth creating another filter for this? It is in the spirit of per-request resource tracking.
         FileUtil.startRequest();
         try
         {
