+++ wiki/src/org/labkey/wiki/WikiController.java
@@ -58,9 +58,7 @@ import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.settings.AdminConsole;
 import org.labkey.api.settings.AppProps;
-import org.labkey.api.util.ContainerTreeSelected;
-import org.labkey.api.util.DiffMatchPatch;
-import org.labkey.api.util.DiffMatchPatch.Diff;
+import org.labkey.wiki.DiffMatchPatch.Diff;
 import org.labkey.api.util.GUID;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.PageFlowUtil;
